+++ ParticleToy/forces/AngularSpringForce.h
@@ -5,7 +5,7 @@
 
 class AngularSpringForce : public Force{
 public:
-    AngularSpringForce(Particle *p1, Particle *p2, Particle *p3, double dist, double rstAngle, double ks, double kd);
+    AngularSpringForce(Particle *p1, Particle *p2, Particle *p3, double dist, float rstAngle, double ks, double kd);
     void calc_Force() override;
     void draw() override;
 
@@ -16,6 +16,6 @@ private:
     Particle *const m_p3;   // particle 2
     double const m_dist;    // rest length
     float originOrient;
-    double m_rstAngle;
+    float m_rstAngle;
     double const m_ks, m_kd; // spring strength constants
 };
